==> 00_DataPrep.ipynb <==
# Collect a data set and prepare it for ML

This is a synthetic data set based on an ML comp dataset

I have uploaded it into a bucket in the same region as this Sagemaker Studio domain 

https://telco-churn-seoul.s3.ap-northeast-2.amazonaws.com/churn_training.csv


import pandas as pd
import numpy as np


import sagemaker
import boto3
from sagemaker import get_execution_role

region = boto3.Session().region_name

session = sagemaker.Session()

# WE NEED A PLACE TO STORE INTERMEDIATE DATA TO BE USED BY ML PROCESSES
# You can modify the following to use a bucket of your choosing
bucket = session.default_bucket()
prefix = 'sagemaker/telco-churn'

role = get_execution_role()

# This is the client we will use to interact with SageMaker AutoPilot
sm = boto3.Session().client(service_name='sagemaker',region_name=region)

Use the following to load the data from S3 into a pandas dataframe

s3 = boto3.client('s3')
obj = s3.get_object(Bucket='telco-churn-seoul', Key='churn_training.csv')
df = pd.read_csv(io.BytesIO(obj['Body'].read()))

df.head()

# Process the data
For the sake of simplicity, we are going to process the data to have numeric only fields.

def process_df(df):
    # change the target col
    df["Churn"] = np.where(df["Churn"] == "Yes", 1, 0)
    # remove unnecessary cols
    df_model_data = df.drop(["CustomerID", "date"], axis=1)
    # convert categories to indicators
    df_model_data = pd.get_dummies(df_model_data)
    return df_model_data

proc = process_df(df)

proc.head()

# Partition the data

Break it into some partitions for training, validation and scoring

target = "Churn"

from sklearn.model_selection import train_test_split
train_prop = 0.6
valid_prop = 0.2
valid_remainder = valid_prop/(1-train_prop)

train, rest = train_test_split(proc, train_size=train_prop, stratify=df[target])
valid, test = train_test_split(rest, train_size=valid_remainder, stratify=rest[target])

# Write to disk

!mkdir data

# WRITE OUT PARTITIONED DATA 
output_path="data"
train.to_csv(output_path + "/train.csv", index=False, header=True)
valid.to_csv(output_path + "/validation.csv", index=False, header=True)
test.to_csv(output_path + "/test.csv", index=False, header=True)
