==> 05_Text_Features.ipynb <==
# Text Feature Processing

In this example we are going to augment a dataset with NLP derived text features using the Amazon Comprehend APIs.

Note: That to execute this Notebook your role will need the following policy attached: ComprehendFullAccess

!wget -O data/smsspamcollection.zip https://www.dt.fee.unicamp.br/~tiago/smsspamcollection/smsspamcollection.zip

!ls -la data

!unzip data/smsspamcollection.zip -d data

!head data/SMSSpamCollection.txt

import pandas as pd
import numpy as np

df = pd.read_csv("data/SMSSpamCollection.txt", delimiter="\t", header=None)
df.columns = ["class","text"]

df.head()

import sagemaker
import boto3

boto_session = boto3.Session()
region = boto_session.region_name
comprehend = boto3.client('comprehend', region_name=region)

# Comprehend API for single block of text

In these examples we use comprehend to extract features from a single block of text.

sample_tweet="It’s always a great day when I can randomly put my equestrian knowledge to good use at work! #AWS #BePeculiar"   

# Key phrases
phrases = comprehend.detect_key_phrases(Text=sample_tweet, LanguageCode='en')

# Entities
entities = comprehend.detect_entities(Text=sample_tweet, LanguageCode='en')

#Sentiments
sentiments = comprehend.detect_sentiment(Text=sample_tweet, LanguageCode='en')


# Print the phrases:
print('------- phrases ---------')
for i in range(0, len(phrases['KeyPhrases'])):
    print((phrases['KeyPhrases'][i]['Text']))
    

# Print the entities with entitity type:
print('------- entity : entity type ---------')
for i in range(0, len(entities['Entities'])):
    print(entities['Entities'][i]['Text'] + ' : ' + entities['Entities'][i]['Type'] )
    
# Print the sentiment:
print('------- sentiment ---------')
print(sentiments['Sentiment'])

## Language detection

It is also possible to first detect the language and then use that in subsequent comprehend API calls for additinal NLP features

lan = comprehend.detect_dominant_language(Text=sample_tweet)

language = lan['Languages'][0]['LanguageCode']

language

# Batch Jobs

Comprehend has separate APIs to score batches of text blocks

https://docs.aws.amazon.com/comprehend/latest/dg/API_BatchDetectSentiment.html

rez = comprehend.batch_detect_sentiment(TextList=['happy days', 'not feeling good'], LanguageCode='en')

rez

# NLP Feature Engineering Pipeline

Now that we understand the fundamentals of the Comprehend APIs, lets build a function that uses it to add features to a dataframe.


# The Comprehend batching scoring API - comprehend.batch_detect_sentiment -
# will only takes a max of 25 documents
# So we build a simple wrapper to score larger numbers of documents


import math

flatten = lambda t: [item for sublist in t for item in sublist]

def batch_sentiment(text_list, lan='en'):
    max_len = 25
    list_len = len(text_list)
    chunks = math.floor(list_len/max_len) + 1
    rez = []
    for index in range(0,chunks):
        start_index = index * max_len
        sublist = text_list[(start_index):(start_index+max_len)]
        xents = comprehend.batch_detect_sentiment(TextList=sublist, LanguageCode=lan)
        rez.append(xents['ResultList'])
    return flatten(rez)


def augment_data_frame(df, col, lan='en'):
    rez = df.copy()
    str_vals = df[col].values.tolist()
    sent = batch_sentiment( str_vals, lan )
    posvals = [ x['SentimentScore']['Positive'] for x in sent]
    negvals = [ x['SentimentScore']['Negative'] for x in sent]
    mixvals = [ x['SentimentScore']['Mixed'] for x in sent]
    rez["Sentiment_positive"] = posvals
    rez["Sentiment_negative"] = negvals
    rez["Sentiment_mixed"] = mixvals
    return rez


augmented = augment_data_frame(df, 'text')

augmented.head()

# Repeat

The same process can be repeated for the other Comprehend APIs to augement the text with additional features
