==> 09_Custom_Kernel.ipynb <==
# Custom Kernel

For this exercise please refer to the following repository: 

https://github.com/john-hawkins/amazon-sagemaker-kernels

It contains an example of how to build a kernel to support R programmers that will be made available inside Sagemaker Studio.

Please Note: this requires additional IAM permissions for the Sagemaker Studio user and access to the internet or an aretfact repository for building the container. This permissions may be restricted to specific users or for DEV environments only.
